move query metadata generation signal to model file
