+++ posthog/models/insight.py
@@ -10,12 +10,14 @@ from django.utils import timezone
 from django_deprecate_fields import deprecate_field
 from rest_framework.exceptions import ValidationError
 from django.db.models import QuerySet
+from django.db.models.signals import pre_save
 from django.contrib.postgres.indexes import GinIndex
 
 from posthog.logging.timing import timed
 from posthog.models.dashboard import Dashboard
 from posthog.models.file_system.file_system_mixin import FileSystemSyncMixin
 from posthog.models.filters.utils import get_filter
+from posthog.models.signals import mutable_receiver
 from posthog.models.utils import sane_repr
 from posthog.utils import absolute_uri, generate_cache_key, generate_short_id
 from posthog.models.file_system.file_system_representation import FileSystemRepresentation
@@ -304,3 +306,17 @@ def generate_insight_filters_hash(insight: Insight, dashboard: Optional[Dashboar
             exc_info=True,
         )
         raise
+
+
+@mutable_receiver(pre_save, sender=Insight)
+def generate_insight_query_metadata_pre_save(sender, instance: Insight, **kwargs):
+    try:
+        instance.generate_query_metadata()
+    except Exception as e:
+        # log and ignore the error, as this is not critical
+        logger.exception(
+            "Failed to generate query metadata for insight",
+            insight_id=instance.id,
+            error=str(e),
+        )
+        capture_exception(e)
